dialog box glitch fixed, more buttons updated, minor refactoring
