+++ app/src/main/AndroidManifest.xml
@@ -35,7 +35,7 @@
         <activity
             android:name=".UI.main_menu.subscription.ExpSubscriptionActivity"
             android:label="My Subscriptions" />
-        <activity android:name=".UI.main_menu.forum.ForumRepliesActivity" android:label="Forum Replies"
+        <activity android:name=".UI.main_menu.specific_experiment_details.discussion.ForumRepliesActivity" android:label="Forum Replies"
             android:theme="@style/EditProfileTheme"
             />
         <activity
