+++ app/src/main/res/layout/activity_forum_replies.xml
@@ -51,7 +51,9 @@
         android:layout_height="wrap_content"
         android:layout_marginRight="20dp"
         android:layout_marginBottom="20dp"
-        android:text="+ REPLY"
+        android:text="+  REPLY"
+        android:backgroundTint="@color/ice_blue"
+
         app:layout_constraintBottom_toBottomOf="@+id/linearLayout3"
         app:layout_constraintEnd_toEndOf="parent"
         />
