+++ app/src/main/res/values/colors.xml
@@ -16,6 +16,7 @@
     <color name="ice_blue">#4dc6f4</color>
 
     <color name="rich_navy_blue">#242747</color>
+    <color name="darker_navy_blue">#171945</color>
     <color name="rich_light_blue">#1B1F5E</color>
 
     <color name="pink">#EC008C</color>
