+++ javadocs/com/example/appraisal/UI/main_menu/forum/ForumRepliesActivity.html
@@ -126,7 +126,7 @@ var activeTableTab = "activeTableTab";
 <li><a href="../../../../../../com/example/appraisal/UI/main_menu/MainMenuCommonActivity.html" title="class in com.example.appraisal.UI.main_menu">com.example.appraisal.UI.main_menu.MainMenuCommonActivity</a></li>
 <li>
 <ul class="inheritance">
-<li>com.example.appraisal.UI.main_menu.forum.ForumRepliesActivity</li>
+<li>com.example.appraisal.UI.main_menu.specific_experiment_details.discussion.ForumRepliesActivity</li>
 </ul>
 </li>
 </ul>
